Allow FillFormatter for individual LineDataSets.
