+++ MPChartExample/src/com/xxmassdeveloper/mpchartexample/LineChartActivity2.java
@@ -29,6 +29,7 @@ import com.github.mikephil.charting.data.filter.Approximator.ApproximatorType;
 import com.github.mikephil.charting.listener.OnChartValueSelectedListener;
 import com.github.mikephil.charting.utils.ColorTemplate;
 import com.github.mikephil.charting.highlight.Highlight;
+import com.xxmassdeveloper.mpchartexample.custom.MyFillFormatter;
 import com.xxmassdeveloper.mpchartexample.notimportant.DemoBase;
 
 import java.util.ArrayList;
@@ -292,6 +293,7 @@ public class LineChartActivity2 extends DemoBase implements OnSeekBarChangeListe
         set1.setFillColor(ColorTemplate.getHoloBlue());
         set1.setHighLightColor(Color.rgb(244, 117, 117));
         set1.setDrawCircleHole(false);
+        //set1.setFillFormatter(new MyFillFormatter(0f));
 //        set1.setDrawHorizontalHighlightIndicator(false);
 //        set1.setVisible(false);
 //        set1.setCircleHoleColor(Color.WHITE);
@@ -317,6 +319,7 @@ public class LineChartActivity2 extends DemoBase implements OnSeekBarChangeListe
         set2.setFillColor(Color.RED);
         set2.setDrawCircleHole(false);
         set2.setHighLightColor(Color.rgb(244, 117, 117));
+        //set2.setFillFormatter(new MyFillFormatter(900f));
 
         ArrayList<LineDataSet> dataSets = new ArrayList<LineDataSet>();
         dataSets.add(set2);
