+++ MPChartLib/src/com/github/mikephil/charting/interfaces/LineDataProvider.java
@@ -1,6 +1,8 @@
 
 package com.github.mikephil.charting.interfaces;
 
+import com.github.mikephil.charting.components.AxisBase;
+import com.github.mikephil.charting.components.YAxis;
 import com.github.mikephil.charting.data.LineData;
 import com.github.mikephil.charting.utils.FillFormatter;
 
@@ -8,18 +10,5 @@ public interface LineDataProvider extends BarLineScatterCandleBubbleDataProvider
 
     public LineData getLineData();
 
-    /**
-     * Sets a custom FillFormatter to the chart that handles the position of the
-     * filled-line for each DataSet. Set this to null to use the default logic.
-     * 
-     * @param formatter
-     */
-    public void setFillFormatter(FillFormatter formatter);
-
-    /**
-     * Returns the FillFormatter that handles the position of the filled-line.
-     * 
-     * @return
-     */
-    public FillFormatter getFillFormatter();
+    public YAxis getAxis(YAxis.AxisDependency dependency);
 }
